server/requirements.txt
Django>=3.2
coverage
livekit-api
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
channels==4.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
Django==5.1.6
django-cors-headers==4.7.0
django-crontab==0.7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dotenv==0.9.9
exceptiongroup==1.2.2
frozenlist==1.5.0
groq==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
kombu==5.4.2
livekit-api==0.8.2
livekit-protocol==0.9.0
multidict==6.1.0
pip==22.0.2
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
setuptools==59.6.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
types-protobuf==4.25.0.20240417
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3

